feat(main): add move_all_files and add empty needed functions

+++ smart_files_gr_13.py
@@ -1,4 +1,5 @@
 import os
+# TODO: fix docs
 
 def word_frequency(string: str) -> dict[str, int]:
     """
@@ -14,45 +15,57 @@ def word_frequency(string: str) -> dict[str, int]:
         ret[word] = ret.get(word, 0) + 1
     return ret
 
-def move_file(fname: str, category: str) -> None:
+def move_file(fname: str, category: str, dir: str = 'archive') -> None:
     """
     Moves a file from the unsorted directory to the sorted directory into the specified category.
     :param fname: the file to move from the unsorted folder
     :param category: the category name to move the file to
+    :param dir: the working directory of sorting, 'archive' by default
+    :raises ValueError: if ``dir`` does not exist in ``.``
     :raises ValueError: if ``fname`` does not exist in ``./archive/unsorted``
     :return:
     """
-    path_to_file = f'./archive/unsorted/{fname}'
-    path_to_category = f'./archive/sorted/{category}'
+    path_to_file = f'./{dir}/unsorted/{fname}'
+    path_to_category = f'./{dir}/sorted/{category}'
     path_to_new_file = f'{path_to_category}/{fname}'
     if not os.path.exists(path_to_file):
         raise ValueError('File to move does not exist.')
+    if not os.path.exists(f'./{dir}'):
+        raise ValueError('Working directory does not exist.')
     if not os.path.exists(path_to_category):
         os.mkdir(path_to_category)
     os.rename(path_to_file, path_to_new_file)
 
+def move_all(info: dict[str, str], dir: str = 'archive') -> None:
+    """
+    Moves all files from ./archive/unsorted into ./archive/sorted by using info in ``info`` dict.
+    :param info: the dictionary to be used as a reference for sorting. 'file_name': 'category_name'
+    :param dir: the working directory of sorting, 'archive' by default
+    :raises ValueError: if ``dir`` does not exist in ``.``
+    :return:
+    """
+    if not os.path.exists(f'./{dir}'):
+        raise ValueError('Working directory does not exist.')
+    for file, category in info:
+        move_file(file, category)
 
 
-# def sort_one_file(fpath: str, sort_dir: str):
-#     """
-#
-#     :param fpath: a string of path leading to the file to be sorted
-#     :param sort_dir: a string of path leading to the directory that contains directories to sort the file to
-#     :raises ValueError: if ``path`` or ``sort_dir`` does not exist
-#     :return: ...
-#     """
-#     if not os.path.exists(fpath):
-#         raise ValueError('Path does not exist')
-#     file = open(fpath, 'r')
-#     content = file.read()
-#     file.close()
-#     word_count = word_frequency(content)
-#     if word_count['test'] and word_count['test'] >= 3:
-#
-#     return ...
-#
-#
-# res = sort_one_file('archive/unsorted/test.txt')
-# print(res)
+def smart_sort_files(path: str) -> None:
+    """
+    Sorts all files by categories into different folders from ./(path)/unsorted/ into ./(path)/sorted
+    :param path: the working directory of the sorting.
+    :raises ValueError: if ``path`` does not exist in ``.``.
+    :return:
+    """
+    ...
+
+def check_accuracy(path: str) -> None:
+    """
+    Checks accuracy of an already performed sorting in ./(path)/sorted by using ./(path)/labels.txt as a reference.
+    :param path: the working directory of previously performed sorting.
+    :raises ValueError: if ``path`` does not exist in ``.``.
+    :return:
+    """
+    ...
 
 move_file('test.txt', 'test')
